Extract DropIn SDK logic from SDK.
