+++ Drop-In/src/test/java/com/braintreepayments/api/DropInClientUnitTest.java
@@ -79,7 +79,7 @@ public class DropInClientUnitTest {
 
     @Test
     public void internalConstructor_setsBraintreeClientWithSessionId() {
-        DropInClient sut = new DropInClient(ApplicationProvider.getApplicationContext(), Fixtures.TOKENIZATION_KEY, "session-id", new DropInRequest());
+        DropInClient sut = new DropInClient(ApplicationProvider.getApplicationContext(), Fixtures.TOKENIZATION_KEY, new DropInRequest());
         assertEquals("session-id", sut.braintreeClient.getSessionId());
     }
